Relocate additional locked-figures files.
