+++ packages/perseus-editor/src/components/graph-locked-figures/color-select.tsx
