+++ packages/perseus-editor/src/components/graph-locked-figures/color-swatch.tsx
