+++ packages/perseus-editor/src/components/graph-locked-figures/coordinate-pair-input.tsx
