+++ packages/perseus-editor/src/components/graph-locked-figures/ellipse-swatch.tsx
