+++ packages/perseus-editor/src/components/graph-locked-figures/labeled-row.tsx
