+++ packages/perseus-editor/src/components/graph-locked-figures/labeled-switch.tsx
