+++ packages/perseus-editor/src/components/graph-locked-figures/line-swatch.tsx
