+++ packages/perseus-editor/src/components/graph-locked-figures/locked-ellipse-settings.tsx
@@ -8,10 +8,10 @@ import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
 import AngleInput from "../angle-input";
-import ColorSelect from "../color-select";
-import CoordinatePairInput from "../coordinate-pair-input";
-import EllipseSwatch from "../ellipse-swatch";
 
+import ColorSelect from "./color-select";
+import CoordinatePairInput from "./coordinate-pair-input";
+import EllipseSwatch from "./ellipse-swatch";
 import LockedFigureSettingsAccordion from "./locked-figure-settings-accordion";
 import LockedFigureSettingsActions from "./locked-figure-settings-actions";
