+++ packages/perseus-editor/src/components/graph-locked-figures/locked-line-settings.tsx
@@ -13,10 +13,9 @@ import {LabelMedium, LabelLarge} from "@khanacademy/wonder-blocks-typography";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import ColorSelect from "../color-select";
-import DefiningPointSettings from "../defining-point-settings";
-import LineSwatch from "../line-swatch";
-
+import ColorSelect from "./color-select";
+import DefiningPointSettings from "./defining-point-settings";
+import LineSwatch from "./line-swatch";
 import LockedFigureSettingsAccordion from "./locked-figure-settings-accordion";
 import LockedFigureSettingsActions from "./locked-figure-settings-actions";
