+++ packages/perseus-editor/src/components/graph-locked-figures/locked-point-settings.tsx
@@ -11,11 +11,10 @@ import {LabelLarge} from "@khanacademy/wonder-blocks-typography";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import ColorSelect from "../color-select";
-import ColorSwatch from "../color-swatch";
-import CoordinatePairInput from "../coordinate-pair-input";
-import LabeledSwitch from "../labeled-switch";
-
+import ColorSelect from "./color-select";
+import ColorSwatch from "./color-swatch";
+import CoordinatePairInput from "./coordinate-pair-input";
+import LabeledSwitch from "./labeled-switch";
 import LockedFigureSettingsAccordion from "./locked-figure-settings-accordion";
 import LockedFigureSettingsActions from "./locked-figure-settings-actions";
