+++ packages/perseus-editor/src/components/graph-locked-figures/polygon-swatch.tsx
