+++ packages/perseus-editor/src/components/interactive-graph-settings.tsx
@@ -17,7 +17,7 @@ import {css, StyleSheet} from "aphrodite";
 import * as React from "react";
 import _ from "underscore";
 
-import LabeledRow from "./labeled-row";
+import LabeledRow from "./graph-locked-figures/labeled-row";
 import ToggleableCaret from "./toggleable-caret";
 
 import type {PerseusImageBackground} from "@khanacademy/perseus";
